Added a toString to GemFireRelation
The toString() is required for GemFireRelation plans to appear properly in GUI else
the default toString() for a case class includes all its arguments which includes all
the column names (e.g. for wide-table it will show a huge box)

Removed explicit "isDebugEnabled" check which was causing a compilation issue
and is not required since Spark's logDebug is a pass-by-name that will
be evaluated only if debug is enabled else the argument will remain as a closure.